pledge component is completed and correctly passes the pledge amount to the new-user component
