+++ client/src/components/index.js
@@ -1,7 +1,9 @@
 import Login from './login/login';
 import NewUser from './newUser/newUser';
+import Pledge from './pledge/pledge'
 
 export {
   Login,
-  NewUser
+  NewUser,
+  Pledge,
 };
