+++ client/src/components/landing.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import './landing.css';
-import Pledge from './pledge.js';
 import { Link } from 'react-router-dom';
 // i removed the handleChange function because it's superfluous. 
 // The buttons will just directly link to the appropriate components instead.
