--- client/src/components/pledge.css
@@ -1,27 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Raleway');
-
-.center{
-	font-family: 'Raleway', sans-serif;
-	background-color: #1abc9c;
-	position: absolute;
-	height: 100%;
-	width: 100%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-.button{
-	font-family: 'Raleway', sans-serif;
-	height: 35px;
-	color: #fff;
-	background-color: #1abc9c;
-	border: none;
-	letter-spacing: 0.2rem;
-	transition: 0.3s opacity ease;
-	cursor: pointer;
-}
-
-.button:hover{
-	opacity: 0.6;
-}
